[widget] fix gap in y axis for vices
This was the result of the position of the container only being stored when
the widget first loads, which means its was vulnerable to changes in
position for the rest of the page as more content loaded. By getting the
position of the container every time the mouse moves, we avoid this issue.